requirements.txt
pymongo==4.6.2
streamlit==1.40.2
streamlit-extras==0.4.0
streamlit-option-menu==0.3.12
extra-streamlit-components==0.1.70
ldap3==2.9.1
python-ldap==3.4.4
openpyxl==3.1.5
